Middleware

+++ routes/web.php
@@ -42,13 +42,13 @@ Route::get('/posts/{post}/edit', 'PostController@edit')->name('post.edit');
 Route::patch('/posts/{post}', 'PostController@update')->name('post.update');
 Route::delete('/posts/{post}', 'PostController@destroy')->name('post.destroy');
 
-Route::get('/admin', 'AdminController@index')->name('admin.index');
-Route::get('/admin/create', 'AdminController@create')->name('admin.post.create');
-Route::post('/admin', 'AdminController@store')->name('admin.post.store');
-Route::get('/admin/{post}', 'AdminController@show')->name('admin.post.show');
-Route::get('/admin/{post}/edit', 'AdminController@edit')->name('admin.post.edit');
-Route::patch('/admin/{post}', 'AdminController@update')->name('admin.post.update');
-Route::delete('/admin/{post}', 'AdminController@destroy')->name('admin.post.destroy');
+Route::get('/admin', 'AdminController@index')->middleware('admin')->name('admin.index');
+Route::get('/admin/create', 'AdminController@create')->middleware('admin')->name('admin.post.create');
+Route::post('/admin', 'AdminController@store')->middleware('admin')->name('admin.post.store');
+Route::get('/admin/{post}', 'AdminController@show')->middleware('admin')->name('admin.post.show');
+Route::get('/admin/{post}/edit', 'AdminController@edit')->middleware('admin')->name('admin.post.edit');
+Route::patch('/admin/{post}', 'AdminController@update')->middleware('admin')->name('admin.post.update');
+Route::delete('/admin/{post}', 'AdminController@destroy')->middleware('admin')->name('admin.post.destroy');
 
 Route::get('/posts/delete','PostController@delete');
 Route::get('/posts/restore','PostController@revdelete');
